refactor(some-filter): split content theming into theme-apply / theme-detector
Pure structural refactor, no behavior change. Decompose the content-script
theming logic into clear modules along the read/write seam:

- color.ts          shared color primitives (parseColor, relativeLuminance,
                    effectiveBgLuminance) — pure, no DOM mutation
- theme-detector.ts read-only classification behind detect(); wraps the
                    existing classifyPage()
- theme-apply.ts    all DOM writes: dark-theme CSS + luminance patcher +
                    legacy filter, behind applyTheme() / restoreVendor()

content.ts shrinks to orchestration (state machine + message wiring) and
now consumes detect()/applyTheme()/restoreVendor(). Legacy-filter and
dark-theme helpers moved out of content.ts.

Tests re-pointed to the new module paths and split to match boundaries
(color.test.ts, theme-detector.test.ts, theme-apply.test.ts), with added
coverage for detector read-only purity and the new apply/restore surface.
67 tests pass; typecheck and content build clean.

Part of #231. Closes #233.
