+++ extensions/some-filter/src/lib/content/theme-apply.ts
@@ -1,24 +1,29 @@
 /**
+ * Theme applier — owns every DOM write involved in theming a page.
  *
- * Two-part dark theme application:
+ * Two strategies live here behind one surface:
+ *   - dark theme: a static CSS layer (buildDarkThemeCSS) plus a JS luminance
+ *     patcher (patchAll + MutationObserver) that tags vendor backgrounds.
+ *   - legacy filter: a single global root `filter: invert(...)` rule.
+ *
+ * Public surface:
+ *   - applyTheme(mode, config?) — apply the dark theme or the legacy filter.
+ *   - restoreVendor()           — remove all theming, returning to native styles.
+ * The granular dark-theme exports (injectDarkTheme/removeDarkTheme/repatchPage)
+ * are retained for the SPA re-patch path and unit tests.
  *
  * Part A — CSS layer (static rules):
- *   Previously scoped to #__sw_page_layer. Now scoped to body (and html),
- *   with :not([data-my-ext]) / :not([data-my-ext] *) guards on rules that
- *   could bleed into extension-owned subtrees.
- *
- *   The [data-my-ext] attribute marks extension-owned nodes. Any node
- *   carrying it — or descended from one — is excluded from theming.
+ *   Scoped to body (and html) with :not([data-my-ext]) / :not([data-my-ext] *)
+ *   guards on rules that could bleed into extension-owned subtrees.
  *
  * Part B — JS luminance patcher (dynamic):
- *   Unchanged in contract. Walks document.body, skips [data-my-ext] nodes
- *   and their descendants via shouldSkip().
- *
- * The #__sw_page_layer wrapper div has been removed entirely. Vendor DOM
- * is left in place; extension nodes self-exclude via the attribute.
+ *   Walks document.body, skips [data-my-ext] nodes and their descendants via
+ *   shouldSkip(). Vendor DOM is left in place; extension nodes self-exclude.
  */
 
-import { parseColor, relativeLuminance } from "./classify"
+import type { FilterConfig } from "@filter/types/config"
+
+import { parseColor, relativeLuminance } from "./color"
 
 export const DARK_THEME_ATTR = "data-sw-dark"
 
@@ -313,7 +318,7 @@ function stopPatchObserver(): void {
   patchObserver = null
 }
 
-// ── Injection ─────────────────────────────────────────────────────────────────
+// ── Dark theme injection ────────────────────────────────────────────────────────
 
 const STYLE_ID = "__sw_dark_theme"
 
@@ -354,4 +359,84 @@ export function removeDarkTheme(): void {
  */
 export function repatchPage(): void {
   patchAll(document.body)
+}
+
+// ── Legacy filter ─────────────────────────────────────────────────────────────
+
+const LEGACY_FILTER_STYLE_ID = "__sw_legacy_filter"
+
+function buildFilterString(config: FilterConfig): string {
+  const parts: Array<string> = []
+
+  if (config.invert !== undefined) parts.push(`invert(${config.invert})`)
+  if (config.hueRotate !== undefined)
+    parts.push(`hue-rotate(${config.hueRotate}deg)`)
+  if (config.sepia !== undefined) parts.push(`sepia(${config.sepia})`)
+  if (config.brightness !== undefined)
+    parts.push(`brightness(${config.brightness})`)
+  if (config.contrast !== undefined) parts.push(`contrast(${config.contrast})`)
+
+  return parts.join(" ")
+}
+
+function applyLegacyFilter(config: FilterConfig): void {
+  let style = document.getElementById(LEGACY_FILTER_STYLE_ID)
+
+  if (!style) {
+    style = document.createElement("style")
+    style.id = LEGACY_FILTER_STYLE_ID
+
+    const root = document.head
+    root.appendChild(style)
+  }
+
+  style.textContent = `
+    html { filter: ${buildFilterString(config)} !important; }
+    img, video, canvas, picture {
+      filter: invert(1) hue-rotate(180deg) !important;
+    }
+  `
+}
+
+function removeLegacyFilter(): void {
+  document.getElementById(LEGACY_FILTER_STYLE_ID)?.remove()
+}
+
+// ── Dark theme activation (internal) ────────────────────────────────────────────
+
+function activateDarkTheme(): void {
+  document.documentElement.setAttribute(DARK_THEME_ATTR, "")
+  injectDarkTheme()
+}
+
+function deactivateDarkTheme(): void {
+  document.documentElement.removeAttribute(DARK_THEME_ATTR)
+  removeDarkTheme()
+}
+
+// ── Public surface ──────────────────────────────────────────────────────────────
+
+export type ThemeMode = "dark" | "legacy"
+
+/**
+ * Apply a theme. `dark` injects the dark-theme CSS layer + patcher; `legacy`
+ * installs the global invert filter (pass `config` for legacy — omitted/undefined
+ * is a no-op). Does not clear the other mode — the orchestrator calls
+ * restoreVendor() first when switching.
+ */
+export function applyTheme(mode: ThemeMode, config?: FilterConfig): void {
+  if (mode === "legacy") {
+    if (config !== undefined) applyLegacyFilter(config)
+    return
+  }
+  activateDarkTheme()
+}
+
+/**
+ * Remove all theming and return the page to its native vendor styles.
+ * Veil teardown remains the orchestrator's responsibility.
+ */
+export function restoreVendor(): void {
+  deactivateDarkTheme()
+  removeLegacyFilter()
 }
